pyproject.toml
[tool.poetry]
name = "marketmaker"
version = "1.0.0"
description = "Suite of tools for managing and creating attack plans"
authors = ["2XXE <root@example.com>"]
readme = "README.md"
# https://spdx.org/licenses/
license = "AGPL-3.0-only"
packages = [
    { include = "libmm" }
]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.10.7"
"ruamel.yaml" = "^0.17.21"
click = "^8.1.3"
stix2 = "^3.0.1"
# TODO: split these into extras based on extension
graphviz = "^0.20.1"
sqlmodel = "^0.0.8"
pysigma = "^0.11.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mm-cli = "libmm.scripts.mmcli:main"
